Transfered user and voice endpoints
