+++ clamor/utils/parse.py
@@ -1,3 +1,5 @@
+from typing import Optional
+
 import re
 
 EMOJI_REGEX = re.compile(r'<a?(:\w+:\d+)>')
@@ -9,3 +11,19 @@ def parse_emoji(emoji: str) -> str:
         emoji = match.group(1)
 
     return emoji
+
+
+def check_username(username: str) -> Optional[str]:
+    if not username:
+        return None
+
+    if 2 > len(username) > 32:
+        raise ValueError('Usernames must be beween 2 and 32 characters long')
+
+    if username in ('discordtag', 'everyone', 'here'):
+        raise ValueError('Restricted username')
+
+    if any(c in ('@', '#', ':', '```') for c in username):
+        raise ValueError('Usernames must not contain "@", "#", ":" or "```"')
+
+    return username.strip()
